+++ trunk/Server/Stump.Server.WorldServer/Game/Effects/Handlers/Spells/Marks/TrapSpawn.cs
@@ -39,8 +39,8 @@ namespace Stump.Server.WorldServer.Game.Effects.Handlers.Spells.Marks
 
             // todo : find usage of Dice.Value
             var trap = EffectZone.ShapeType == SpellShapeEnum.Q ?
-                new Trap((short)Fight.PopNextTriggerId(), Caster, Spell, Dice, trapSpell, TargetedCell, GameActionMarkCellsTypeEnum.CELLS_CROSS, Effect.ZoneSize) :
-                new Trap((short)Fight.PopNextTriggerId(), Caster, Spell, Dice, trapSpell, TargetedCell, Effect.ZoneSize);
+                new Trap((short)Fight.PopNextTriggerId(), Caster, Spell, Dice, trapSpell, TargetedCell, GameActionMarkCellsTypeEnum.CELLS_CROSS, (byte) Effect.ZoneSize) :
+                new Trap((short)Fight.PopNextTriggerId(), Caster, Spell, Dice, trapSpell, TargetedCell, (byte) Effect.ZoneSize);
 
             Fight.AddTriger(trap);
             return true;
